Update polish #177 #58 #166
